# [Erlang] sublime-syntax format version 2