@deathaxe 2020-09-09
Makes sense as the performance drop is nearly the same with all syntaxes I have made benchmarks for. Only my new Java syntax which heavily makes use of branch-points to support case insensitive object types etc. runs quite a bit more slower than before.

Just wanted to be sure everything is working as expected and we don't miss something. Most files are small enough, so users might not even notice it.