@wbond 2020-09-09
Erf, or maybe a useless loop over every named context was added to a hot spot in the lexing engine… 🤦‍♂️