@wbond 2020-09-10
Funny story, but I removed that debug loop and was getting about 25% better performance…